adds estimated time, total size, and pagination check
desc: 3 major feature introduction: a. adding estimated time to complete the record based on total records and the time taken to download records
b. adding total request size parameter to help people gauge the size of their query
c. added check in query if 'id' is present or not. if 'id' is not present then we go for paginating directly. If they mistakenly left the id parameter we will throw a warning that id is missing, if you mistakenly left it, please include it explicitely